backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
certifi==2026.2.25
click==8.3.1
colorama==0.4.6
fastapi==0.135.1
greenlet==3.3.2
groq>=0.11.0
gunicorn>=21.2.0
h11==0.16.0
idna==3.11
pinecone-client==6.0.0
pinecone-plugin-interface==0.0.7
psycopg2-binary==2.9.11
pydantic==2.12.5
pydantic-settings>=2.0.0
pydantic_core==2.41.5
python-dateutil==2.9.0.post0
python-dotenv>=1.0.0
python-multipart==0.0.22
redis==7.3.0
six==1.17.0
SQLAlchemy==2.0.48
sqlmodel==0.0.37
starlette==0.52.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.42.0
